requirements.txt
Django==5.0
pika==1.3.1
djangorestframework==3.14.0
python-dotenv==1.0.0
